backend/requirements.txt
Flask>=2.3.2,<2.4
flasgger==0.9.7.1
flask_sqlalchemy==3.1.1
pytest==8.3.2
flask_bcrypt==1.0.1
flask_jwt_extended==4.6.0
flask_cors==5.0.0
python-dotenv==1.0.1
